plugin/file: move wildcard additional test into wildcard_test.go
Requested in review: keep the wildcard tests in one file.
